# Feature/optimize hash