--- a/Project.toml
+++ b/Project.toml
@@ -1,7 +1,7 @@
 name = "NMEAParser"
 uuid = "e33883a1-8aa8-4e6e-a42d-0daf9e203b40"
 authors = ["Nicholas Shindler <nick@shindler.tech>"]
-version = "2.1.0"
+version = "2.1.1"
 
 [weakdeps]
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"