# Optimize the hash function for checksum validation

added an internal `_char_xor` function that handles Char to UInt8 conversion and allows calling `foldl` directly on the string.

This gives significant performance improvement:

#### old version
```julia_repl
julia> @benchmark NMEAParser.hash_msg(message)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  268.443 μs …  5.542 ms  ┊ GC (min … max): 0.00% … 91.39%
 Time  (median):     284.955 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   291.835 μs ± 75.036 μs  ┊ GC (mean ± σ):  0.33% ±  1.30%

  ▅▂▁▁▂█▅▄▄▄▆▆▅▄▄▄▄▄▃▃▄▃▃▃▃▂▂▂▂▁▁▁▁▁▁▁  ▁                      ▂
  ███████████████████████████████████████████▇█▇▇█▇▇▆▇▇▆▄▆▆▆▆▆ █
  268 μs        Histogram: log(frequency) by time       359 μs <

 Memory estimate: 12.47 KiB, allocs estimate: 281.
 
julia> @benchmark NMEAParser.nmea_parse(message)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  276.243 μs …   8.496 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     286.042 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   294.879 μs ± 125.454 μs  ┊ GC (mean ± σ):  0.48% ± 1.60%

  ▅▇    █                                                        
  ██▄▂▂▄█▇▃▂▂▄▆▅▂▂▂▅▄▃▂▂▃▂▃▂▂▂▁▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  276 μs           Histogram: frequency by time          360 μs <

 Memory estimate: 14.06 KiB, allocs estimate: 290.
```
 #### new version
 ```julia_repl
julia> @benchmark NMEAParser.hash_msg(message)
BenchmarkTools.Trial: 10000 samples with 651 evaluations.
 Range (min … max):  188.210 ns … 320.198 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     188.425 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   196.120 ns ±  14.939 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █    ▆    ▁ ▁▂▁          ▁▂▂▂  ▂ ▁▁                   ▁▁      ▁
  █▆▆▄▆█▆▄▅▆█▆███████▇▆▆▄▆▄███████████▆▇▆▆███▇▇▆▆▅▆▅▁▄▇███▇▆▆▇▅ █
  188 ns        Histogram: log(frequency) by time        250 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
 
julia> @benchmark NMEAParser.nmea_parse(message)
BenchmarkTools.Trial: 10000 samples with 8 evaluations.
 Range (min … max):  3.911 μs … 256.883 μs  ┊ GC (min … max): 0.00% … 94.44%
 Time  (median):     4.138 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.494 μs ±   3.500 μs  ┊ GC (mean ± σ):  1.03% ±  1.33%

   ▅█▅▅                                                        
  ▂████▆▂▁▁▁▁▁▁▃▃▃▄▃▃▂▃▃▃▃▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  3.91 μs         Histogram: frequency by time        7.01 μs <

 Memory estimate: 1.59 KiB, allocs estimate: 9.
 ```
 
For all the benchmarks `message = raw"$GPGGA,135020.000,5540.2676,N,01231.2830,E,2,10,0.8,24.7,M,41.5,M,3.8,0000"`
 is used as the test data.